Add support for ST7567 lcd displays
These displays can be obtained very cheaply on places like aliexpress and ebay and use a similar addressing method to the SSD1306. Always wanted to use this library with them so I added an initialization script for it.

How to use it:

oled.begin(&ST7567_128x64, 0x3F);

Everything just works!